+++ release/linux/scripts/build-4-nodes-chain.sh
@@ -112,9 +112,6 @@ function four() {
 
 ################################################# Main #################################################
 function main() {
-    assignDefault
-    readParam
-
     four
 }